A Xamarin Android build! Hooray!
This is using the shared code from the PCL ReactiveFlickr project and
compiles fine. There's a warning thrown in Xamarin Studio about the
Android project not referencing Microsoft.Bcl.Build, but it doesn't need
to because Mono takes care of that for us.

The Xamarin Android project is just a shell for now, it needs to be hooked
up.